+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v3.1.1
+* Added: Graph.getExtendedData to get a single extended data row
 * Fixed: InMemory update extended data with different value
 
 # v3.1.0